Detect and fill holes in weather data.
Optional hole size (in minutes) parameter triggers an attempt to find holes in the weather data.
Found holes are filled with data from Ambient if possible.
After holes filled with data from Ambient, another hole scan is performed, and detected holes are filled with NULLs for the data, and every 5 min for dateutc and date.